feat: Show a sample of Findings in the Welcome message
