+++ packages/cli/src/cmds/inventory-report/MarkdownReport.ts
@@ -1,9 +1,11 @@
 import Handlebars from 'handlebars';
-import { Report } from '../inventory/Report';
 import { join } from 'path';
 import { readFile } from 'fs/promises';
 import assert from 'assert';
 import { existsSync } from 'fs';
+
+import { Report } from '../inventory/Report';
+import helpers from '../compare-report/helpers';
 
 const TemplateDirectory = [
   '../../../resources/inventory-report', // As packaged
@@ -31,6 +33,7 @@ export default async function generateReport(
       reportTemplate = new WelcomeReport();
       break;
   }
+
   return reportTemplate.generateReport(report);
 }
 
@@ -67,30 +70,6 @@ export class ReportTemplate {
   }
 
   static helpers(): { [name: string]: Function } | undefined {
-    const inspect = (value: any) => {
-      return new Handlebars.SafeString(JSON.stringify(value, null, 2));
-    };
-
-    const length = (...list: any[]): number => {
-      const _fn = list.pop();
-      let result = 0;
-      for (const item of list) {
-        if (Array.isArray(item)) {
-          result += item.length;
-        } else if (item.constructor === Map) {
-          result += item.size;
-        } else if (typeof item === 'object') {
-          result += Object.keys(item).length;
-        }
-      }
-      return result;
-    };
-
-    const coalesce = (...list: any[]): number => {
-      const _fn = list.pop();
-      return list.find((item) => item !== undefined && item !== '');
-    };
-
     const select_values_by_key_range = (
       data: Record<number, number>,
       keyMin: number,
@@ -103,7 +82,6 @@ export class ReportTemplate {
         .map((k) => data[k]);
       return result;
     };
-    const extractArrayValue = (args: any[]): any[] => (Array.isArray(args[0]) ? args[0] : args);
 
     const sum_values_by_key_range = (
       data: Record<number, number>,
@@ -114,20 +92,10 @@ export class ReportTemplate {
       return values.reduce((a, b) => a + (b || 0), 0);
     };
 
-    const sum = () => {
-      const args = [...arguments];
-      const _fn = args.pop();
-      const values = extractArrayValue(args);
-      return values.reduce((a, b) => a + (b || 0), 0);
-    };
-
     return {
-      coalesce,
-      inspect,
-      length,
       select_values_by_key_range,
-      sum,
       sum_values_by_key_range,
+      ...helpers,
     };
   }
