+++ packages/cli/src/cmds/inventory-report/inventoryReport.ts
@@ -6,6 +6,7 @@ import generateReport from './MarkdownReport';
 import { warn } from 'console';
 import { cwd } from 'process';
 import { join } from 'path';
+import { Report } from '../inventory/Report';
 
 export const command = 'inventory-report <report-json-file> [output-file]';
 export const describe = 'Generate a report document describing the current state of a repository.';
@@ -47,7 +48,7 @@ export const handler = async (argv: any) => {
 
   const { reportJsonFile, outputFile, templateName } = argv;
 
-  const report = JSON.parse(await readFile(reportJsonFile, 'utf-8'));
+  const report: Report = JSON.parse(await readFile(reportJsonFile, 'utf-8'));
 
   const reportMD = await generateReport(templateName, report);
   if (outputFile) {
